[Fix] Forgot to create the mountpoint directory before starting FUSE system... [Improvement] Allow server to run on interface assigned by installation script instead of only "tapipop".

+++ scripts/make.sh
@@ -1,11 +1,22 @@
 #!/bin/bash
 # This script builds the fushare project and copies binaries to make a package.
 
+print_help_and_exit() {
+  printf "Usage: %s: [-bvc] \n" $(basename $0) >&2; 
+  cat <<EOF
+  -b Build application.
+  -v Verbose.
+  -c Compress and make a tarball.
+EOF
+  exit 2
+}
+
 while getopts "bvc" optNames; do
 case "$optNames" in
   b) build=1;;
   v) verbose=1;;
   c) compress=1;;
+  ?) print_help_and_exit;;
   esac
 done
 
@@ -19,17 +30,22 @@ fushare_src="$fushare_proj/src"
 if [ "$build" ]; then
   echo "Building Fushare..."
   "$fushare_scripts/make_fushare.sh" -b
+  if [ $? -ne 0 ]; then
+    echo "Build failed. Exiting..."
+    exit 1
+  fi
+
+  client_proj_name="FushareApp"
+  web_proj_name="Fushare.Web"
+  client_dir="$fushare_src/$client_proj_name/bin/l4n"
+  web_dir="$fushare_src/$web_proj_name"
+
+  rsync -avLm --delete --include-from="$scripts_dir/rsync.rules" --exclude-from="$scripts_dir/rsync.rules" "$client_dir/" "$proj_dir/client/bin/"
+  rsync -avLm --delete --include-from="$scripts_dir/rsync.rules" --exclude-from="$scripts_dir/rsync.rules" "$web_dir/" "$proj_dir/server/bin/"
 fi
 
-client_proj_name="FushareApp"
-web_proj_name="Fushare.Web"
-client_dir="$fushare_src/$client_proj_name/bin/l4n"
-web_dir="$fushare_src/$web_proj_name"
-
-rsync -avLm --delete --include-from="$scripts_dir/rsync.rules" --exclude-from="$scripts_dir/rsync.rules" "$client_dir/" "$proj_dir/client/bin/"
-rsync -avLm --delete --include-from="$scripts_dir/rsync.rules" --exclude-from="$scripts_dir/rsync.rules" "$web_dir/" "$proj_dir/server/bin/"
-
 if [ "$compress" ]; then
-  tar czvf "$proj_dir/../gatorrent.tar.gz" "$proj_dir" --exclude=".git*"
+  # Make a tarball gatorrent.tar.gz with gatorrent as the root directory inside.
+  tar -C $proj_dir/.. -czvf "$proj_dir/../gatorrent.tar.gz" gatorrent --exclude=".git*" --exclude="*~"
 fi
